dora_node_hub/dora_gemini_diff_drive_navigation/pyproject.toml
[project]
name = "dora_gemini_diff_drive_navigation"
version = "0.1.0"
authors = [{ name = "Franco Cipollone", email = "franco.c@ekumenlabs.com" }]
description = "dora_gemini_diff_drive_navigation"
license = { text = "BSD 3-Clause" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "dora-rs >= 0.3.9",
  "google-genai >= 1.23.0",
  "numpy >= 1.26.4",
  "opencv-python >= 4.1.1",
]

[dependency-groups]
dev = [
  "pytest >=8.1.1",
  "ruff >=0.9.1",
]

[project.scripts]
dora_gemini_diff_drive_navigation = "dora_gemini_diff_drive_navigation.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
